requirements.txt
paho-mqtt==2.0.0
sseclient==0.0.27
crccheck==1.3.0